source/backend/tripplanner/requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-cors-headers==3.1.1
django-dotenv==1.4.2
django-filter==2.2.0
djangorestframework==3.10.3
idna==2.8
Markdown==3.1.1
oauthlib==3.1.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.6
